% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoiatestpackage.R
\docType{package}
\name{Sequoiatestpackage-package}
\alias{Sequoiatestpackage}
\alias{Sequoiatestpackage-package}
\title{Sequoiatestpackage: this package runs foo and ggplot2 for quick test set up}
\description{
This could be longer to develop this
main problem it solves.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
